consolidate and format function names
